divide to different components
